+++ buildSrc/src/main/kotlin/LibraryConfig.kt
@@ -2,7 +2,7 @@ import org.gradle.api.JavaVersion
 
 object LibraryConfig {
     const val Group = "io.github.pubiqq"
-    const val Version = "1.2.0-SNAPSHOT"
+    const val Version = "2.0.0-SNAPSHOT"
 
     const val MinSdk = 14
     const val CompileSdk = 33