+++ lifecycleprops/src/main/kotlin/com/pubiqq/lifecycleprops/LifecycleAwareInitializationStrategy.kt
@@ -1,27 +1,33 @@
 package com.pubiqq.lifecycleprops
 
-import androidx.lifecycle.Lifecycle
-
 /**
- * Specifies the initialization strategy for the property value.
+ * Specifies the strategy to initialize the property.
  */
-@LifecycleAwareOptions
+@LifecycleAwareConfigurationApi
 public enum class LifecycleAwareInitializationStrategy {
     /**
      * The property initializer will be called immediately when the delegate is initialized.
      */
-    Eager,
+    OnInit,
 
     /**
-     * The property initializer will be called the first time the property is accessed.
-     * If the initializer throws an exception, it will attempt to reinitialize the value
-     * at next access.
+     * The property initializer will be called the first time the property is accessed _explicitly_.
      *
-     * Note that handling lifecycle methods also requires accessing the property.
-     * It follows that, for example, if we access the property value when
-     * [Lifecycle.Event.ON_RESUME] event is reached, but we also have [Lifecycle.Event.ON_START]
-     * handler, the initializer will be called when the [Lifecycle.Event.ON_START] event
-     * is reached.
+     * The initializer will not be called if the property access is initiated by the lifecycle event
+     * handler when trying to get the receiver value. The behavior in this case is determined by the
+     * [LifecycleAwareReadWriteConfiguration.allowSkipHandlerAccessToUninitializedProperty] option.
+     *
+     * If the initializer throws an exception, it will attempt to reinitialize the property value at
+     * next access.
      */
-    Lazy
+    OnPropertyAccess,
+
+    /**
+     * The property initializer will be called the first time the property is accessed, including if
+     * it's initiated by the lifecycle event handler to get the receiver value.
+     *
+     * If the initializer throws an exception, it will attempt to reinitialize the property value at
+     * next access.
+     */
+    OnAnyAccess
 }