+++ lifecycleprops/src/main/kotlin/com/pubiqq/lifecycleprops/LifecycleAwareReadOnlyOptions.kt
@@ -1,7 +1,0 @@
-package com.pubiqq.lifecycleprops
-
-@LifecycleAwareOptions
-public class LifecycleAwareReadOnlyOptions(
-    public val initializationStrategy: LifecycleAwareInitializationStrategy,
-    public val deinitializationStrategy: LifecycleAwareDeinitializationStrategy
-)