+++ lifecycleprops/src/main/kotlin/com/pubiqq/lifecycleprops/LifecycleAwareReadWriteOptions.kt
@@ -1,6 +1,0 @@
-package com.pubiqq.lifecycleprops
-
-@LifecycleAwareOptions
-public class LifecycleAwareReadWriteOptions(
-    public val deinitializationStrategy: LifecycleAwareDeinitializationStrategy
-)