+++ lifecycleprops/src/main/kotlin/com/pubiqq/lifecycleprops/internal/ReadWritePropertyLifecycleObserver.kt
@@ -4,10 +4,10 @@ import androidx.lifecycle.DefaultLifecycleObserver
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleEventObserver
 import androidx.lifecycle.LifecycleOwner
-import com.pubiqq.lifecycleprops.LifecycleAwareDeinitializationStrategy
+import com.pubiqq.lifecycleprops.LifecycleAwareReadWriteConfiguration
 
 internal class ReadWritePropertyLifecycleObserver<T : Any>(
-    private val deinitializationStrategy: LifecycleAwareDeinitializationStrategy,
+    private val configuration: LifecycleAwareReadWriteConfiguration<T>,
     private val onCreate: (T.() -> Unit)? = null,
     private val onStart: (T.() -> Unit)? = null,
     private val onResume: (T.() -> Unit)? = null,
@@ -18,58 +18,86 @@ internal class ReadWritePropertyLifecycleObserver<T : Any>(
 ) : DefaultLifecycleObserver, LifecycleEventObserver {
 
     private var _value: T? = null
+        set(value) {
+            val oldValue = field
+            if (oldValue !== value && oldValue != null) {
+                configuration.onClear(oldValue)
+            }
+
+            field = value
+        }
 
     var value: T
-        get() = _value ?: throw IllegalStateException("Value is not available")
+        get() = _value ?: error("The property is not initialized")
         set(value) {
-            _value = value
+            @Suppress("LiftReturnOrAssignment")
+            if (_value == null) {
+                _value = value
+            } else {
+                if (configuration.allowReassign) {
+                    _value = value
+                } else {
+                    error("The property cannot be reassigned")
+                }
+            }
+        }
+
+    private val valueForHandlers: T?
+        get() {
+            _value?.let { return it }
+
+            return if (configuration.allowSkipHandlerAccessToUninitializedProperty) {
+                null
+            } else {
+                error("The property is not initialized")
+            }
         }
 
     override fun onCreate(owner: LifecycleOwner) {
         onCreate?.let { onCreate ->
-            _value?.let { onCreate(it) }
+            valueForHandlers?.let { value -> onCreate(value) }
         }
     }
 
     override fun onStart(owner: LifecycleOwner) {
         onStart?.let { onStart ->
-            _value?.let { onStart(it) }
+            valueForHandlers?.let { value -> onStart(value) }
         }
     }
 
     override fun onResume(owner: LifecycleOwner) {
         onResume?.let { onResume ->
-            _value?.let { onResume(it) }
+            valueForHandlers?.let { value -> onResume(value) }
         }
     }
 
     override fun onPause(owner: LifecycleOwner) {
         onPause?.let { onPause ->
-            _value?.let { onPause(it) }
+            valueForHandlers?.let { value -> onPause(value) }
         }
     }
 
     override fun onStop(owner: LifecycleOwner) {
         onStop?.let { onStop ->
-            _value?.let { onStop(it) }
+            valueForHandlers?.let { value -> onStop(value) }
         }
     }
 
     override fun onDestroy(owner: LifecycleOwner) {
         onDestroy?.let { onDestroy ->
-            _value?.let { onDestroy(it) }
+            valueForHandlers?.let { value -> onDestroy(value) }
         }
     }
 
     override fun onStateChanged(source: LifecycleOwner, event: Lifecycle.Event) {
         onAny?.let { onAny ->
-            _value?.let { onAny(it, event) }
+            valueForHandlers?.let { value -> onAny(value, event) }
         }
 
-        if (event == Lifecycle.Event.ON_DESTROY
-            && deinitializationStrategy == LifecycleAwareDeinitializationStrategy.OnDestroy
-        ) {
-            _value = null
+        if (event == Lifecycle.Event.ON_DESTROY) {
+            if (configuration.shouldNullOutTheProperty) {
+                _value = null
+            }
         }
     }
 }