+++ lifecycleprops/src/main/kotlin/com/pubiqq/lifecycleprops/internal/utils/Close.kt
@@ -1,0 +1,9 @@
+package com.pubiqq.lifecycleprops.internal.utils
+
+import android.os.Build
+
+internal fun <T> T.closeIfPossible() {
+    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT && this is AutoCloseable) {
+        close()
+    }
+}